+++ dimensions/1/src/io.c
@@ -2,13 +2,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-double* read(char *src, int *L){
-    double *data;
+long double* read(char *src, int *L){
+    long double *data;
     FILE *in;
     int it;
     in = fopen(src, "r");
     fscanf(in, "%d", L);
-    data = (double*) calloc(*L, 2*sizeof(double));
+    data = (long double*) calloc(*L, 2*sizeof(long double));
     for(it=0; it < *L; it++){
         fscanf(in, "%lf", data+2*it);
         data[2*it+1]=0;
@@ -17,7 +17,7 @@ double* read(char *src, int *L){
     return data;
 }
 
-void write(char *dst, int L, double *dft){
+void write(char *dst, int L, long double *dft){
     FILE* out;
     int it;
     out = fopen(dst, "w");
